Removed obsolite connectors and integrated proper functions into Hassan's connector :)
