+++ src/connections/connector.py
@@ -9,6 +9,10 @@ class Connector:
         "Y": 0.0,
         "Z": 0.0,
     }
+    lastPosL = 0
+    lastPosR = 0
+    l = 0
+    r = 0
 
     def __new__(cls):
         if cls._instance is None:
@@ -43,39 +47,57 @@ class Connector:
         line = self.serialConnection.readline().decode("utf-8").strip()
         if line:
             logging.debug(line)
-        return line
 
-    def write_data(self, data):
-        self.serialConnection.write(data.encode('utf-8'))
+        match line.split(',')[0].strip():
+            case "CAPTURE":
+                return line
+            case "ENCODER":
+                self.parse_encoder(line)
+                return "ENCODER"
+            case "GYRO":
+                self.parse_gyro(line)
+                return "GYRO"
+            case _:
+                return line
+
+    def parse_encoder(self, line):
+        if line != ['']:
+            data = list(map(int, line))
+            self.l = data[1] - self.lastPosL
+            self.r = data[2] - self.lastPosR
+            self.lastPosL = data[1]
+            self.lastPosR = data[2]
+
+    def parse_gyro(self, line):
+        pass
 
     def get_gyro_data(self):
         return self.gyro
 
-    def send_command(self, cmd):
+    def write_data(self, cmd):
         self.serialConnection.write(cmd.encode('utf-8'))
-        time.sleep(0.2)
 
     def forward(self):
-        self.send_command("w")
+        self.write_data("w")
 
     def backward(self):
-        self.send_command("s")
+        self.write_data("s")
 
     def left(self):
-        self.send_command("a")
+        self.write_data("a")
         print('Left function')
 
     def right(self):
-        self.send_command("d")
+        self.write_data("d")
     
     def start(self):
-        self.send_command("z")
+        self.write_data("z")
 
     def stop(self):
-        self.send_command("x")
+        self.write_data("x")
 
     def drive_autonomously(self):
-        self.send_command("t")
+        self.write_data("t")
 
     def drive_manually(self):
-        self.send_command("m")
+        self.write_data("m")
